feat: create routes for 'checkout' and 'home'
