+++ package.json
@@ -13,7 +13,9 @@
     "@phosphor-icons/react": "^2.1.7",
     "phosphor": "^0.7.1",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router": "^7.1.1",
+    "react-router-dom": "^7.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
