+++ src/Router.tsx
@@ -1,0 +1,13 @@
+import { Routes, Route } from 'react-router-dom';
+
+import { Home } from './pages/Home';
+import { Checkout } from './pages/Checkout';
+
+export function Router() {
+  return (
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/checkout" element={<Checkout />} />
+    </Routes>
+  );
+}
