+++ src/pages/Checkout.tsx
@@ -1,0 +1,3 @@
+export function Checkout() {
+  return <div>Checkout</div>;
+}
